// backend/src/integrations/m3_integration/co_header/model/co-header.interface.ts
/* eslint-disable prettier/prettier */
export interface coheader {
  CO_ID: string;
  FG_Number: string;
  Schedule_ID: string;
  BLI_Del_Date: any;
  CPO_Del_Date: any;
  Buyer_PO: string;
  Buyer_Styleno: string;
  Range: string;
  Item_SKU_Code: string;
  Colour_Code:string;
  Colour_Description: string;
  Size_Code: string;
  Size_Description: string;
  ZFeature_Code: string;
  ZFeature_Description: string;
  MO_Number: string;
  Quantity: number;
  MO_Status: number;
  SMV: number;
  Product_Category: string;
  Buyer_Name: string;
  Club_Number: number;
  MKR_Status: boolean;
  PED: any;
  Extra: number;
}